% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensg_to_gene_symbol.R
\name{ensg_to_ensp}
\alias{ensg_to_ensp}
\title{ensg_to_ensp Function}
\usage{
ensg_to_ensp(ensg_character_vector)
}
\arguments{
\item{ensg_character_vector}{A character vector of ensemble gene ids (such as 'ENSG00000134057')}
}
\description{
This function allows you to Gets biomart table (august/2018)
 to change ensgs to ensp
}
