% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_candidates_from_network.R
\name{get_candidates_from_network}
\alias{get_candidates_from_network}
\title{get_candidates_from_network Function}
\usage{
get_candidates_from_network(g, network_measure, blacklist)
}
\arguments{
\item{g}{a igraph}

\item{network_measure}{either 'strength' or 'eigen_centrality'}

\item{blacklist}{vector of genes to be removed from candidate list, presumably because once they were degs}
}
\description{
get_candidates_from_network Function
}
