% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_encode_edges.R
\name{get_encode_edges}
\alias{get_encode_edges}
\title{get_encode_edges Function}
\usage{
get_encode_edges(vector_of_gene_symbols, type_of_regulation, verbose = T)
}
\arguments{
\item{vector_of_gene_symbols}{A character vector of gene symbols}

\item{type_of_regulation}{Either 'direct' or 'common_tf'}

\item{verbose}{If true, adds verbosity. Defaults to true.}
}
\description{
get_encode_edges Function
}
