% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_humap_edges.R
\name{get_humap_edges}
\alias{get_humap_edges}
\title{get_humap_edges Function}
\usage{
get_humap_edges(vector_of_gene_symbols)
}
\arguments{
\item{vector_of_gene_symbols}{A character vector of gene symbols}
}
\description{
get_humap_edges Function
}
