% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_KEA_edges.R
\name{get_KEA_edges}
\alias{get_KEA_edges}
\title{get_KEA_edges Function}
\usage{
get_KEA_edges(vector_of_gene_symbols)
}
\arguments{
\item{vector_of_gene_symbols}{A character vector of gene symbols}
}
\description{
get_KEA_edges Function
}
