% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_disgenet_info.R
\name{list_disgenet_info}
\alias{list_disgenet_info}
\title{list_disgenet_info Function
list diseases for a given query}
\usage{
list_disgenet_info(disease, database_path)
}
\arguments{
\item{disease}{string containing the disease you are looking for (use first letter uppercase )}

\item{database_path}{The addres of the .db file downloaded from
http://www.disgenet.org/ds/DisGeNET/files/current/disgenet_2017.db.gz}
}
\description{
list_disgenet_info Function
list diseases for a given query
}
\examples{
url <- 'http://www.disgenet.org/ds/DisGeNET/files/current/disgenet_2017.db.gz'
download.file(url,'./disgenet.R')
unzip file manually to data directory

disgenet_dengue <- get_disgenet_info('Dengue', database_path = 'data/disgenet_2017.db')
}
