% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_candidates_from_network.R
\name{get_candidates_from_network}
\alias{get_candidates_from_network}
\title{get_candidates_from_network Function}
\usage{
get_candidates_from_network(g, network_measure, top_candidates, blacklist)
}
\arguments{
\item{network_measure}{either 'strength' or 'eigen_centrality'}

\item{top_candidates}{the portion of candidates that will be taken}
}
\description{
get_candidates_from_network Function
}
